fix(removing excess components): removing excess components
