+++ package.json
@@ -8,6 +8,8 @@
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
+    "@types/classnames": "^2.3.1",
+    "@types/styled-components": "^5.1.13",
     "classnames": "^2.3.1",
     "mapbox-gl": "^2.4.1",
     "node-sass": "^6.0.1",
@@ -22,6 +24,7 @@
     "redux-mock-store": "^1.5.4",
     "regenerator-runtime": "^0.13.9",
     "sass-loader": "^12.1.0",
+    "styled-components": "^5.3.1",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
