+++ src/LoftTaxi.jsx
@@ -4,7 +4,7 @@ import React from "react";
 import { MapWithAuth } from "./Components/Map/Map";
 import propTypes from "prop-types";
 import { connect } from "react-redux";
-import RouteCollection from "./Components/Login/PrivateRoutes";
+import PrivateRoutes from "./Components/PrivateRoutes/PrivateRoutes";
 
 function App(props) {
   // if (window.localStorage.getItem("view") != "map") {
@@ -14,14 +14,14 @@ function App(props) {
               <main data-testid = "main">
                   <LoginPage/>
               </main>
-              <RouteCollection />
+              <PrivateRoutes />
           </>
         );
     } else {
       return <><main data-testid = "map-wrapper">
                   <MapWithAuth {...props} />
              </main>
-             <RouteCollection /></>
+             <PrivateRoutes /></>
     }
 }
